+++ packages/svelte/src/lib/ConvexNudge.svelte
@@ -18,6 +18,7 @@
   } from '@convex-nudge/core';
 
   export let className = DEFAULT_OPTIONS.className;
+  export let link: NudgeOptions['link'] = undefined;
   export let referralCode = DEFAULT_OPTIONS.referralCode;
   export let logoSize = DEFAULT_OPTIONS.logoSize;
   export let textSize = DEFAULT_OPTIONS.textSize;
@@ -84,7 +85,7 @@
     <div style={styleToString(containerStyles)}>
       <div style={styleToString(lineStyles)} />
       <a
-        href={getConvexUrl(referralCode)}
+        href={getConvexUrl(link || referralCode)}
         target="_blank"
         rel="noopener noreferrer"
         style={styleToString(LINK_STYLES)}